rename & move helper_bot code
refactor helper bot files
