+++ src/workers/mod.rs
@@ -1,6 +1,5 @@
 pub mod async_race_worker;
 pub mod forfeit_own_races;
-pub mod race_event_status_worker;
 pub mod racetime_scanner_worker;
 #[cfg(feature = "racetime_bot")]
 pub mod upcoming_races_worker;
